[v0.88][maintainability] Split identity_cmd.rs by responsibility (Closes #1826)
